perf(module): rename all eth to ethereum
