+++ src/chain/ethereum/mod.rs
@@ -1,19 +1,19 @@
 //! Ethereum types
 // mod confirmation;
-mod ethash_proof;
+mod ethash;
 mod header;
-mod header_thing;
-mod mmr_proof;
+mod mmr;
+mod parcel;
 mod receipt;
 mod runtime;
 
 pub use self::{
-    ethash_proof::{EthashProof, EthashProofJson},
+    ethash::{EthashProof, EthashProofJson},
     header::{EthereumHeader, EthereumHeaderJson, EthereumHeaderRPC},
-    header_thing::{
+    mmr::{HeaderStuff, HeaderStuffJson, MMRProof, MMRProofJson},
+    parcel::{
         EthereumHeaderThing, EthereumHeaderThingJson, EthereumHeaderThingWithConfirmationJson,
     },
-    mmr_proof::{HeaderStuff, HeaderStuffJson, MMRProof, MMRProofJson},
     receipt::{
         EthereumReceiptProof, EthereumReceiptProofJson, EthereumReceiptProofThing,
         EthereumReceiptProofThingJson, RedeemFor,
